requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
pydantic>=2.11.0
python-multipart>=0.0.9
httpx>=0.28.0
pytesseract>=0.3.10
Pillow>=10.4.0
google-cloud-firestore>=2.21.0
firebase-admin>=6.6.0
pytest>=8.3.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
google-genai>=1.0.0
python-dotenv>=1.0.0
